split witness and allocator
