+++ src/witness.rs
@@ -4,81 +4,11 @@ extern "C" {
     pub fn wasm_witness_pop() -> u64;
 }
 
+use crate::allocator::{get_cursor_ptr, start_alloc_witness, stop_alloc_witness};
+use crate::allocator::{MAX_WITNESS_OBJ_SIZE, WITNESS_AREA, WITNESS_AREA_END};
 use crate::require;
 use primitive_types::U256;
-use std::alloc::{GlobalAlloc, Layout, System};
 use std::mem::size_of;
-use std::ptr::null_mut;
-
-static mut WITNESS_AREA: usize = 0;
-static mut WITNESS_AREA_END: usize = 0;
-const MAX_WITNESS_OBJ_SIZE: usize = 40 * 1024;
-const MAX_SUPPORTED_ALIGN: usize = 4096;
-
-static mut ALLOC_WITNESS: bool = false;
-static mut SIMPLE_ALLOCATOR: SimpleAllocator = SimpleAllocator {
-    area: [0; MAX_WITNESS_OBJ_SIZE],
-    remaining: MAX_WITNESS_OBJ_SIZE,
-};
-
-/// SETUP for WITNESS_AREA and WITNESS_AREA_END
-pub fn init_simple_allocator() -> *mut u8 {
-    unsafe {
-        let base_addr = SIMPLE_ALLOCATOR.area.as_ptr();
-        WITNESS_AREA = base_addr as usize;
-        WITNESS_AREA_END = WITNESS_AREA + MAX_WITNESS_OBJ_SIZE;
-        base_addr as *mut u8
-    }
-}
-
-struct SimpleAllocator {
-    pub area: [u8; MAX_WITNESS_OBJ_SIZE],
-    remaining: usize,
-}
-
-struct HybridAllocator {}
-
-unsafe fn start_alloc_witness() {
-    ALLOC_WITNESS = true;
-}
-
-unsafe fn stop_alloc_witness() {
-    ALLOC_WITNESS = false;
-    SIMPLE_ALLOCATOR.remaining = MAX_WITNESS_OBJ_SIZE;
-}
-
-unsafe impl Sync for HybridAllocator {}
-
-#[global_allocator]
-static ALLOCATOR: HybridAllocator = HybridAllocator {};
-
-unsafe impl GlobalAlloc for HybridAllocator {
-    unsafe fn alloc(&self, layout: Layout) -> *mut u8 {
-        if ALLOC_WITNESS {
-            let size = layout.size();
-            let align = layout.align();
-            let align_mask_to_round_down = !(align - 1);
-            if align > MAX_SUPPORTED_ALIGN {
-                return null_mut();
-            }
-            if size > SIMPLE_ALLOCATOR.remaining {
-                return null_mut();
-            }
-            SIMPLE_ALLOCATOR.remaining -= size;
-            SIMPLE_ALLOCATOR.remaining &= align_mask_to_round_down;
-            let area_ptr = SIMPLE_ALLOCATOR.area.as_ptr() as *mut u8;
-            area_ptr.add(SIMPLE_ALLOCATOR.remaining)
-        } else {
-            System.alloc(layout)
-        }
-    }
-    unsafe fn dealloc(&self, ptr: *mut u8, layout: Layout) {
-        if ALLOC_WITNESS {
-        } else {
-            System.dealloc(ptr, layout);
-        }
-    }
-}
 
 pub trait WitnessObjWriter {
     fn to_witness(&self, ori_base: *const u8, wit_base: *const u8);
@@ -228,10 +158,7 @@ fn prepare_witness_obj<Obj: Clone + WitnessObjReader + WitnessObjWriter, T>(
 ) -> () {
     let b = gen(t);
     let c = Box::new(b.clone());
-    let ori_base = unsafe {
-        let area = SIMPLE_ALLOCATOR.area.as_ptr() as *mut u8;
-        area.add(SIMPLE_ALLOCATOR.remaining)
-    };
+    let ori_base = get_cursor_ptr();
     unsafe {
         wasm_witness_insert((c.as_ref() as *const Obj as *const u8).sub_ptr(ori_base) as u64);
     }
@@ -275,11 +202,11 @@ fn load_witness_obj<Obj: Clone + WitnessObjReader + WitnessObjWriter, T>(
 
 #[cfg(feature = "zktest")]
 pub(crate) mod test {
-    use super::init_simple_allocator;
     use super::load_witness_obj;
     use super::prepare_witness_obj;
     use super::WitnessObjReader;
     use super::WitnessObjWriter;
+    use crate::allocator::init_simple_allocator;
     use crate::require;
     use crate::wasm_dbg;
     use derive_builder::WitnessObj;
